Load config before dispatching command
Update checker's behaviour is defined by the config file. So the configuration loading needs to be moved to the very beginning, before the update is checked.

I defined config.Config instance as a global variable. Thsi way one does not need to pass it as an argument to every function possible.